+++ module-cellular/at/src/URC_CTZE.cpp
@@ -4,44 +4,44 @@
 #include "../URC_CTZE.hpp"
 #include <log/debug.hpp>
 
+#include <module-utils/date/include/date/date.h>
+
 using namespace at::urc;
 
 CTZE::CTZE(const std::string &val) : Any(val)
 {}
 
-auto CTZE::what() -> std::string
+auto CTZE::what() const noexcept -> std::string
 {
     return urc_name;
 }
 
-struct tm CTZE::getTimeInfo(void)
+auto CTZE::isValid() const noexcept -> bool
 {
-    struct tm timeinfo;
-    std::string dateTimeStr(tokens[static_cast<uint32_t>(tokensDescription::date)] + "," +
-                            tokens[static_cast<uint32_t>(tokensDescription::time)]);
-    utils::findAndReplaceAll(dateTimeStr, "\"", "");
+    return tokens.size() == magic_enum::enum_count<Tokens>();
+}
 
-    std::stringstream stream(dateTimeStr);
-    stream >> std::get_time(&timeinfo, "%Y/%m/%d,%H:%M:%S");
-    constexpr uint32_t signLenght  = 1;
-    constexpr uint32_t valueLength = 2;
-    constexpr uint32_t signPos     = 0;
+auto CTZE::getTimeInfo() const noexcept -> tm
+{
+    using namespace std::chrono;
 
-    utils::findAndReplaceAll(tokens[static_cast<uint32_t>(tokensDescription::GMTdiference)], "\"", "");
-    auto sign  = tokens[static_cast<uint32_t>(tokensDescription::GMTdiference)].substr(signPos, signLenght);
-    auto value = tokens[static_cast<uint32_t>(tokensDescription::GMTdiference)].substr(signLenght, valueLength);
+    std::tm timeinfo{};
+    if (isValid()) {
+        std::string dateTimeStr(tokens[Tokens::Date] + "," + tokens[Tokens::Time]);
+        utils::findAndReplaceAll(dateTimeStr, "\"", "");
 
-    uint32_t offset;
-    if (utils::toNumeric(value, offset)) {
-        constexpr uint32_t quatersInHour = 4;
-        offset /= quatersInHour;
-        if (sign == "+") {
-            timeinfo.tm_hour += offset;
-        }
-        else if (sign == "-") {
-            timeinfo.tm_hour -= offset;
-        }
+        std::stringstream stream(dateTimeStr);
+        date::sys_seconds tp;
+        stream >> date::parse("%Y/%m/%d,%H:%M:%S", tp);
+
+        auto gmtDifferenceStr = tokens[Tokens::GMTDifference];
+        utils::findAndReplaceAll(gmtDifferenceStr, "\"", "");
+
+        int gmtDifference                      = utils::getValue<int>(gmtDifferenceStr);
+        constexpr auto minutesInQuartersOfHour = 15;
+        constexpr auto secondsInMinutes        = 60;
+        auto time = system_clock::to_time_t(tp) + gmtDifference * minutesInQuartersOfHour * secondsInMinutes;
+        timeinfo  = *gmtime(&time);
     }
-
     return timeinfo;
 }